QuickForm/requirements.txt
Flask>=2.0.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-Bcrypt>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.5.0
pymysql>=1.1.0